webapp/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
aiosqlite==0.20.0
pydantic[email]==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
websockets==12.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
imbalanced-learn>=0.12.0
joblib>=1.3.2
aiofiles==23.2.1
python-dotenv==1.0.1
